M    core.py * made releases a more real atomic unit. more work might be needed, which   requires thinking, which I can not do right now. * cleanRepo has ignore support. it is a list set in the configuration of path   roots which will not be scanned for unwanted files.

+++ psync/core.py
@@ -15,29 +15,32 @@ import logging
 logger = logging.getLogger('psync.core')
 logger.setLevel(logLevel)
 
-class ProtocolError (Exception):
+class DictException (Exception):
     def __init__ (self, **kwargs):
         self.__dict__.update(kwargs)
         self.args= kwargs
-        # super (ProtocolError, self).__init__ ()
+
+class ProtocolError (DictException):
+    pass
 
 class DependencyError (Exception):
-    def __init__ (self, **kwargs):
-        self.__dict__.update(kwargs)
-        self.args= kwargs
-        # super (ProtocolError, self).__init__ ()
+    pass
 
 class Psync(object):
     def __init__ (self, verbose=False, **kwargs):
-        self.failed= [] # files that failed to download
-        self.keep= {}
-        self.repoFiles= 0
-        self.downloadedSize= 0
         self.verbose= verbose
-        
-        logger.debug (self.__dict__)
         self.__dict__.update(kwargs)
-        logger.debug (self.__dict__)
+        
+        # file tracking
+        self.releaseFailed= False # something has failed for the release
+        self.failedFiles= [] # files that failed to download
+        self.keep= {}
+        
+        # counters
+        self.repoFiles= 0
+        # repo, release or arch?
+        self.repoSize= 0
+        self.downloadedSize= 0
 
     # simulate to be a dict for suitability to % operator
     def __getitem__ (self, key):
@@ -83,12 +86,12 @@ class Psync(object):
             else:
                 if self.verbose:
                     # logger.info ("%s: already here, skipping" % _file)
-		    pass
+                    pass
         self.keep[_file]= 1
 
         if ans==0x1600:
-            self.failed.append (_file)
-            self.aFileHasFailed= True
+            self.failedFiles.append (_file)
+            self.releaseFailed= True
         elif ans==0x02:
             # curl is not here
             raise DependencyError (package='curl')
@@ -113,35 +116,25 @@ class Psync(object):
         for distro in distros:
             # summary.append ("distro: %s" % distro)
             self.distro= distro
-	    try:
-        	for release in self.releases:
-            	    self.release= release
-            	    # releaseSummary= self.processRelease ()
-            	    self.processRelease ()
-	    except Exception, e:
-		logger.info ('processing %(repo)s/%(distro)s/%(release)s failed due to' % self)
-		logger.info (e)
-		print_exc ()
-        	if ( self.debugging or
-	             (isinstance (e, IOError) and e.errno==errno.ENOSPC) or
-		      isinstance (e, KeyboardInterrupt)):
-		    # debugging, out of disk space or keyb int
-		    raise e
-													                   
+            for release in self.releases:
+                self.release= release
+                # releaseSummary= self.processRelease ()
+                self.processRelease ()
+        
         # summary of failed pkgs
-        if self.failed!=[]:
+        if self.failedFiles!=[]:
             logger.info ("failed packages:")
-            for i in self.failed:
+            for i in self.failedFiles:
                 logger.info (i)
-            logger.info ("----- %d/%d package(s) failed" % (len (self.failed), self.repoFiles))
+            logger.info ("----- %d/%d package(s) failed" % (len (self.failedFiles), self.repoFiles))
             # this is gonna be ugly
             # so ugly I sent it to another function
-            self.keepOldFiles ()
         else:
             # summary+= releaseSummary
+            pass
 
-    	    # clean up
-    	    self.cleanRepo ()
+        # clean up
+        self.cleanRepo ()
 
         # return summary
     
@@ -180,32 +173,69 @@ class Psync(object):
         Returns a list of strings with a summary of what was done.
         It is for human consumption.
         """
-        summary= []
-        # we force to be at least one arch!
-        # some third party repos have none
-        # slack doesn't either
-        archs= getattr (self, 'archs', [ None ])
-        for arch in archs:
-            self.arch= arch
-            # msg= "architecture %s:" % arch
-            # summary.append (msg)
-            # summary.append ("~" * len (msg))
-            modules= getattr (self, 'modules', [ None ])
-            for module in modules:
-                # won't log what module we are processing
-                self.module= module
-                # summary+= self.process ()
-                self.process ()
-                
-            # summary.append (("total update: %7.2f MiB" % (self.downloadedSize/1048576.0)).rjust (75))
-            # summary.append ('')
-            # summary.append ('')
+        try:
+            # summary= []
+            self.releaseFailed= False
             
-            # reset count
-            self.downloadedSize= 0
-
+            archs= getattr (self, 'archs', [ None ])
+            for arch in archs:
+                self.arch= arch
+                # msg= "architecture %s:" % arch
+                # summary.append (msg)
+                # summary.append ("~" * len (msg))
+                modules= getattr (self, 'modules', [ None ])
+                for module in modules:
+                    # won't log what module we are processing
+                    self.module= module
+                    # summary+= self.process ()
+                    self.process ()
+                    
+                # summary.append (("total update: %7.2f MiB" % (self.downloadedSize/1048576.0)).rjust (75))
+                # summary.append ('')
+                # summary.append ('')
+                
+                # reset count
+                self.downloadedSize= 0
+    
+        except Exception, e:
+            self.releaseFailed= True
+            logger.info ('processing %(repo)s/%(distro)s/%(release)s failed due to' % self)
+            logger.info (e)
+            print_exc ()
+            if ( self.debugging or
+                    (isinstance (e, IOError) and e.errno==errno.ENOSPC) or
+                    isinstance (e, KeyboardInterrupt)):
+                # debugging, out of disk space or keyb int
+                raise e
+            
+        if self.releaseFailed:
+            self.keepOldReleaseFiles ()
+        
         # return summary
 
+    def keepOldReleaseFiles (self):
+        # force it to use the old databases
+        try:
+            self.tempDir= '.'
+            archs= getattr (self, 'archs', [ None ])
+            for arch in archs:
+                self.arch= arch
+                modules= getattr (self, 'modules', [ None ])
+                for module in modules:
+                    self.module= module
+                    for filename, size in self.files ():
+                        filename= os.path.normpath (("%(repoDir)s/%(baseDir)s/" % self)+filename)
+                        logger.debug ('keeping %s' % filename)
+                        self.keep[filename]= 1
+                    databases= self.finalDBs()
+                    for (database, critic) in databases:
+                        dst= os.path.normpath (("%(repoDir)s/%(baseDir)s/" % self)+database)
+                        self.keep[dst]= 1
+        except IOError:
+            # some database does not exist in the mirror
+            # so, wipe'em all anyways
+            pass
+    
     def process (self):
         """
         Process one module.
@@ -213,7 +243,6 @@ class Psync(object):
         It is for human consumption.
         """
         # summary= []
-        self.aFileHasFailed= False
         try:
             # download databases
             self.baseDir= self.baseDirTemplate % self
@@ -232,18 +261,18 @@ class Psync(object):
             # yes, there is a BUG here, but it's not that grave.
             # which bug? I don't remember :(
             # planned to be fixed in 0.2.5 or never
-            if not self.save_space and not self.dry_run and self.failed==[] and not self.process_old:
+            if not self.save_space and not self.dry_run and self.failedFiles==[] and not self.process_old:
                 self.updateDatabases ()
             else:
                 logger.debug ("save: %s, dry: %s, failed: %s" %
-                    (self.save_space, self.dry_run, self.failed))
+                    (self.save_space, self.dry_run, self.failedFiles))
             
         except Exception, e:
             logger.info ('processing %s failed due to %s' % (self.repo, e))
             if ( self.debugging or
-        	 (isinstance (e, IOError) and e.errno==errno.ENOSPC) or
+                 (isinstance (e, IOError) and e.errno==errno.ENOSPC) or
                  isinstance (e, KeyboardInterrupt) or
-	    	 isinstance (e, ProtocolError) ):
+                 isinstance (e, ProtocolError) ):
                 # debugging, out of disk space or keyb int
                 print_exc ()
                 raise e
@@ -252,13 +281,21 @@ class Psync(object):
 
     def cleanRepo (self):
         logger.debug ('here comes the janitor')
+        ignorePaths= getattr (self, 'ignore', [])
+        logger.debug (ignorePaths)
         for (path, dirs, files) in os.walk (self.repoDir):
-            for _file in files:
-                filepath= os.path.join (path, _file)
-                if not self.keep.has_key (filepath):
-                    # delete de bastard
-                    logger.info ('deleting %s' % filepath)
-                    os.unlink (filepath)
+            ignore= False
+            for ignorePath in ignorePaths:
+                if path.startswith (ignorePath):
+                    ignore= True
+                
+            if not ignore:
+                for _file in files:
+                    filepath= os.path.join (path, _file)
+                    if not self.keep.has_key (filepath):
+                        # delete de bastard
+                        logger.info ('deleting %s' % filepath)
+                        os.unlink (filepath)
         logger.debug ('janitor finished cleaning')
         files= self.keep.keys ()
         files.sort()
@@ -280,7 +317,7 @@ class Psync(object):
             databaseUrl= ("%(repoUrl)s/%(baseDir)s/" % self)+database
     
             found= grab (dababaseFilename, databaseUrl,
-                            limit=self.limit, progress=self.progress, cont=False)
+                         limit=self.limit, progress=self.progress, cont=False)
             if found!=0 and critic:
                 raise ProtocolError (proto=databaseUrl[:databaseUrl.index (':')].upper (), code=found, url=databaseUrl)
